% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProductPrice.R
\name{getProductPrice}
\alias{getProductPrice}
\title{getProdcutPrice()}
\usage{
getProductPrice(
  price = 5.5,
  price_condition_operator = "greater than",
  rating = 2,
  rating_condition_operator = "greater than"
)
}
\arguments{
\item{price}{a positive number.}

\item{price_condition_operator}{the condition you're looking for, 'greater than', 'less than', 'equal'}

\item{rating}{a number between 0 to 5.}

\item{rating_condition_operator}{the condition you're looking for, 'greater than', 'less than', 'equal'}
}
\value{
a dataframe of product and product feature the user queried
}
\description{
This is a function for user to query makeup product list according to product pricing and rating conditions
}
\author{
Ting-An Lai, \email{tl2960@columbia.edu}
}
